Added flatten
