+++ src/editor/mod.rs
@@ -17,6 +17,8 @@ use crate::get_contour_mut;
 mod toggles;
 pub use self::toggles::*;
 
+pub mod export;
+
 pub mod util;
 
 pub mod viewport;
